added the main page
It features an image slideshow at the top with titles and descriptions. The images automatically change, and there are dots at the bottom to indicate the current image. Although a "Get Started" button is included, it's currently hidden. In essence, it's a screen for viewing a rotating set of images with descriptions, and it provides visual cues to show which image is currently displayed. The hidden button may serve a purpose in the app, but it's not active on this screen.